requirements.txt
boto3
python-dotenv
requests
fastapi[standard]
flask
janus
pythonping
speedtest-cli

salad-cloud-sdk